+++ api/src/main/kotlin/dev/kosmx/needle/database/Init.kt
@@ -3,10 +3,10 @@ package dev.kosmx.needle.database
 import dev.kosmx.needle.core.AssetChecker
 import dev.kosmx.needle.core.ClassChecker
 import dev.kosmx.needle.database.hardCodedDetectors.HardCodedDetectors
-import dev.kosmx.needle.logger
 import kotlinx.coroutines.*
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
+import org.slf4j.kotlin.getLogger
 import org.slf4j.kotlin.warn
 import java.io.IOException
 import java.io.InputStream
@@ -22,6 +22,7 @@ import kotlin.streams.asSequence
  * Initialize database, get ready for checking
  */
 object Database {
+    private val logger by getLogger()
 
 
     fun init(databaseUrl: String?, dataPath: Path) {
