+++ api/src/main/resources/logging.properties
@@ -1,3 +1,3 @@
 
-dev.kosmx.needle.lib=WARNING
+dev.kosmx.needle.util=WARNING
 me.coley=WARNING
